# arc: interrupts: start latency measurement early