For some reason we start measurement of latency a little bit late, after quite some instructions are already executed.

Move read_timer_start_of_isr() to the real entry of generic interrupt handler. And with that we do see some difference on `benchmarks/timing_info`.
```
Before: Interrupt latency    :  20 cycles ,    40 ns
After:  Interrupt latency    :  62 cycles ,   124 ns
```