@abrodkin 2019-08-28
> pls consider the case of firq with 2 reg banks, the 2nd reg bank's sp may point to invalid places. 

Good point! But do we handle such a case (with not initialized SP in the second register bank)?

> Also, as sp is changed, stack checking may raise exception.

See my respin. Added support for stack checking even in secure mode.