@vonhust 2019-08-30
@abrodkin  for normal irq,  i am ok, also ok for stack checking. My concern is in firq. Could you run a sanitycheck for nsim_em (with firq and 2 reg banks)  with CONFIG_EXECUTION_BENCHMARKING on ?