@abrodkin 2019-08-30
> 
> 
> @abrodkin for normal irq, i am ok, also ok for stack checking. My concern is in firq. Could you run a sanitycheck for nsim_em (with firq and 2 reg banks) with CONFIG_EXECUTION_BENCHMARKING on ?

Ok I see the problem now. Fast IRQs (FIRQ) are supposed to use exactly the same stack as Regular IRQs (RIRQ). Normally SP is properly set in the first register bank as it's used everywhere except FIRQ where the second bank is used. So on entry to FIRQ the first thing we need to do is to extract SP value from the first bank and set the second bank's SP with that value.

For that to happen we need:
1. Figure out we're handling FIRQ but not RIRQ by checking bit 0 in `IRQ_ACT` (active IRQs) as FIRQs by definition have the highest priority.
2. Switch to the first bank to read current SP.
3. Switch back to the second bank and apply current SP value.

It makes no sense to mess with all that just for a simple timer read.

So let's forget about all that until we do clean-up and simplification of interrupt handling code for ARC.