@ruuddw 2019-08-30
I guess an alternative could be to have optimized (asm), architecture specific timer readout, instead of the generic C implementation that requires a valid C stack pointer/frame to be available. Or, some combination: read the timer and save it in a variable in the ISR, and in the generic C code use that variable.