@abrodkin 2019-08-30
> I guess an alternative could be to have optimized (asm), architecture specific timer readout, instead of the generic C implementation that requires a valid C stack pointer/frame to be available. Or, some combination: read the timer and save it in a variable in the ISR, and in the generic C code use that variable.

Well our implementation doesn't strictly requires stack for timer read-out. But for reading timer value and storing that value to some memory location we obviously need to use 2 general-purpose registers:
1. For keeping value read from Timer0
2. For keeping location of global variable used for delta calculation (where we write read value)

Thus we need to save temporary contents of those 2 regs on stack, otherwise we'll lose their content for good.