@abrodkin 2019-08-31
Ok there's another take and this time much funnier.
The problem we faced with before was a need to save 2 registers before reading value from the timer and writing to the memory location. But what if we may do something that doesn't require us to use any general purpose registers? Like instead of reading initial timer value we may set it with say... zero!

So that's what we may do:
```
diff --git a/arch/arc/core/isr_wrapper.S b/arch/arc/core/isr_wrapper.S
index a8e5a106f0..c31370f421 100644
--- a/arch/arc/core/isr_wrapper.S
+++ b/arch/arc/core/isr_wrapper.S
@@ -202,6 +202,10 @@ From RIRQ:
  */

 SECTION_FUNC(TEXT, _isr_wrapper)
+#ifdef CONFIG_EXECUTION_BENCHMARKING
+       sr 0, [_ARC_V2_TMR0_COUNT]
+#endif
+
 #if defined(CONFIG_ARC_FIRQ)
 #if CONFIG_RGF_NUM_BANKS == 1
 /* free r0 here, use r0 to check whether irq is firq.
@@ -297,9 +301,6 @@ SECTION_FUNC(TEXT, _isr_demux)
        push r59
 #endif

-#ifdef CONFIG_EXECUTION_BENCHMARKING
-       bl read_timer_start_of_isr
-#endif
        /* cannot be done before this point because we must be able to run C */
        /* r0 is available to be stomped here, and exit_tickless_idle uses it */
        exit_tickless_idle
```

And it works quite OK-ish:
```
***** Booting Zephyr OS build v2.0.0-rc1-287-g91c7d0c13aaf *****
starting test - Time Measurement
Timing Results: Clock Frequency: 5 MHz
Context switch                               :  99 cycles , 19800 ns
Interrupt latency                            :  37 cycles ,  7400 ns
...
```

Now a couple of disadvantages:
1. Now for simplicity I zero ARC core Timer0 which is used for normal time keeping and even though `tests/benchmarks/timing_info` still happen to work, in other cases it will lead to unexpected behavior.
2. We may use ARC core Timer1 for this purpose as it's not used for anything in Zephyr yet... but there's no guarantee it exists in the SoC... but so far I haven't heard of any designs lacking Timer1 so possibly that's a very good solution.

@vonhust @ruuddw what do you guys think?