@ruuddw 2019-09-02
When using another timer dedicated for this specific ISR benchmarking, I think the 'set to zero' trick works - in the generic case, however, I'm afraid interrupts could interfere with other benchmark timestamp reads.