@ruuddw 2020-09-10
@abrodkin: with CONFIG_EXECUTION_BENCHMARKING removed, I think this PR is obsolete. Close it?
Instead, we could consider if it makes sense to use the new dedicated benchmark/timestamp timer API with a separate timer/counter. Not sure though if there are real benefits (accuracy, benchmarking overheads).