@abrodkin 2020-09-10
@ruuddw indeed, this one is obsolete now. Though we need to think about some way for characterization of things like that so that we may at least track it and be able to identify regressions on IRQ handling latencies etc.